ACPI: Add copyright notice to ACPI files where code restructurings are made.
Based on the reference ACPI implementation, sdfirm redesigns most
of the table management code to make it adaptive to multi-threading
environment.  Many redundant code and mixed logics are cleaned up
so that the new locking facilities can be put in the right place
to make table management safer for uninstallation.

Copyright notice is added due to the redesigns.

Signed-off-by: Lv Zheng <zetalog@gmail.com>