ACPI: Fix a bug that table count is not calculated correctly.
There is a case FACS is not loaded at boot stage, when it gets
reloaded, the table count is errornously added.
This patch fixes this issue by setting initial table count to 0
and invoking acpi_table_list_acquire() for fixed tables so that
all table counting is managed by single interface.

Signed-off-by: Lv Zheng <zetalog@gmail.com>