ACPI: Fix a bug for the file that does not contain ACPI table.
It is detected that when a file does not contain ACPI table, the
Length field will probably contain a very large number that even
is greater than the file size.  In the original code, it will
stupidly allocate a buffer for this wrong number.  This patch
prevents such wrong allocation from happening.

Signed-off-by: Lv Zheng <zetalog@gmail.com>