ACPI: Fix several issues in the original code.
This commit fixes several issues in the original code.
Such fixes improves winacpi to live in an environment that multiple
threads can install/uninstall same table.

To achieve this, ACPI_TABLE_IS_UNINSTALLING must be the last flag
unset by an uninstallation step and acpi_get_table() is changed to
always fill acpi_table_desc.pointer.  There is a missing "break"
detected in this commit.  acpi_table_decrement() is converted to
be a locked operation and only acpi_os_free() is unlocked inside of
this function.  This commit also includes small fixes in simulator.

Signed-off-by: Lv Zheng <zetalog@gmail.com>