ACPI: Cleanup API locking/unlocking.
This commit cleans up locking/unlocking environment of the APIs, so
that:
1. For APIs do not care about the invocation environment, there is
   no under score prefix (if no conflicts) or "____" under score
   prefix (if conflicts) for the function name.
2. For APIs must be invoked without lock held, there is no under
   score prefix for the function name.
3. For APIs must be invoked with lock held, there is "__" under
   score prefix for the function name.
The this commit carefully checks the implementation of the
functions, unlocks them when necessary.

There is a dead lock detected when acpi_finalize_tables() is
is invoked and parallel installation can increase the table  count
again.  This is also fixed in this commit by retry the table
uninstallations.

Signed-off-by: Lv Zheng <zetalog@gmail.com>