ACPI: Add split validation/invalidation logic.
For external usages, the acpi_get_table()/acpi_put_table() clones
might be re-designed so that when the table is uninstalled, we
don't need to wait for all external usages to run to an end brought
by acpi_put_table().

For internal usages, current logic still need to be kept.  This
patch thus splits and refines the logic to be used for internal
purposes.  The refining includes state machine enhancement and
reference reduced.

This patch also enables table loading code but still disables
parser code.  The reason to do so in the same patch is the table
loading code also need to be refined in the same way as the table
validation code.

Signed-off-by: Lv Zheng <zetalog@gmail.com>