ACPI: Cleanup acpi_get_table()/acpi_put_table() APIs.
This patch introduces new acpi_table structure to hold tables
returned from acpi_get_table() and its clones.  The major change
brought by this patch is there will always be table acquired by
this function so that we do not need to implement wait for the
invalidation operation.

As some OSPMs will have early boot stage where mapped pointer
should not be kept after this stage, if the external usage of a
particular table ends, the invalidation must be performed.  It
would be a trouble if such invalidation occurred during
re-installation where wait must be implemented to reduce the overall
number of table descriptors.
While with all external usages holding table copys, we do not need
to performed such wait as external usages will have nothing to do
with the VALIDATED/INVALIDATED table state and reference count can
also be reduced.

There are also 2 bugs detected by this patch:
1. acpi_emu_load_table() need to be fixed as it may return error
   for acpi_read_table_file.
2. some ACPI_TABLE_SOLVE_INDIRECT() invocations need to be replaced
   with new __acpi_table_solve_indirect() where ddb might be
   invalid.

Signed-off-by: Lv Zheng <zetalog@gmail.com>