ACPI: Add early stage distinguishing support.
This commit adds early stage support for virtual addressed tables.
In this stage, acpi_os_allocate/acpi_os_free may not be available.
In the late stage, it is still required that the virtual addressed
tables are duplicated so that the table unloading process doesn't
need to wait for the completion of invalidation.

Signed-off-by: Lv Zheng <zetalog@gmail.com>