ACPI: Make table reloading wait optional.
If OSPM allows table to be loaded again from its own (from the
definition blocks contained by itself), we shouldn't wait until
uninstallation completes for the reloading.
This commit makes reloading wait optional, the default behaviour
is still wait.  With this option enabled, the number of the table
descriptors will increase to hold the uninstalling tables as well
as the installing/installed new tables.

Signed-off-by: Lv Zheng <zetalog@gmail.com>