ACPI: Fix an issue in extended opcode parsing.
The extended opcode is comprised by a 2-bytes byte stream. This
makes the opcode looking more like a big-endian value as a whole.
So we should stop using ACPI_DECODE16 to parse the opcode.

This patch also hacks all variable type object parsing to return
a simple AMLCode like package for them so that tests are possible.

Signed-off-by: Lv Zheng <zetalog@gmail.com>