ACPI: Allow some opcodes to be executed earlier than parsing their list elements.
This patch adds support to allow some opcodes to be executed ealier
than parsing the whole TermObj. The syntax for such opcodes
normally contains a PkgLength part, before parsing each element,
the header of the package need to be executed. Or we'll see some
the objects created by the header part are forward referenced by
their children.
We add an indication into the acpi_term_cb to support this.

Signed-off-by: Lv Zheng <zetalog@gmail.com>