ACPI: Move scope indication from parser to executer.
The scope information shouldn't belong to the parser state which will be
stacked because of new inner-AML-term inside of an TermArg and only the
executer need to scope information to create new namespace node from the
correct scope.
So this patch moves acpi_namespace_node from acpi_parser to acpi_interp.

Signed-off-by: Lv Zheng <zetalog@gmail.com>