riscv: Cleanup smp entry and scratch/stack calculation
This patch sorts TP/SP calculation functions, and sorts secondary boot
entries. Note that currently, the secondary boot spin uses variable in the
text section, which should be fixed in the future.

This patch also adds useful macros of ABI_PERCPU_STACKxxx to be used for
"RISCV_EXIT_x" priviledge, and unifies stack size using 4K that is the
same size of PAGE.

Due to reimplementation of the entire SBI entry code using completely
different software architectures, removes the old copyrights and
authorship.

Signed-off-by: Lv Zheng <zhenglv@hotmail.com>