bench: Run bench command twice to ensure all CPUs execute
In the first running of "bench sync all test 1 2 1" after an SMP
enabled sdfirm booting over spike -p4 , only CPU 0 executes the test
while other CPUs does not. The reason is not clear at present.  To
make sure that all CPUs execute the test, the second running of the
bench command is added.

- Use two "help" commands in tempalte defconfig
- Replace each "help" commands with certain bench command when
  generating config instance.
- Use larger memory size.

Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>