duowen: Add non-dual socket mode PLIC support
The silicon contains lots of bugs concerning dual socket mode PLIC.
There are several incompatility between the current silicon and the
spec. At this point, no fix is welcome while a special mode of
disabling dual socket support is proposed so that no quirks are
required to perform what can be provided by current buggy state.
This patch adds code to be compatible with the new default PLIC
working mode.

Signed-off-by: Lv Zheng <zhenglv@hotmail.com>